backend/requirements.txt
# ----- FastAPI Core -----
fastapi
uvicorn[standard]
python-multipart
aiofiles
pydantic
sqlalchemy

# ----- Image & YOLO -----
opencv-python
ultralytics
torch
torchvision

# ----- AI / LLM / Web Search -----
openai
requests
tavily-python
python-dotenv
google-generativeai
pandas
 

# ----- Utils -----
pillow
python-dateutil
pytz


# ----- Optional -----
starlette
apscheduler

# ----- RAG -----
chromadb
sentence-transformers
numpy
pypdf